online-recommendation-service/requirements.txt
fastapi~=0.111.0
pydantic~=2.7.1
uvicorn~=0.29.0
requests~=2.31.0
redis==5.2.1
PyYAML==6.0.2
